codigo examenMB modularizado dividido en valores, categoria, subcategoria,singleton
