+++ src/java/ni/gob/minsa/sivipcan/controlador/ExamenEJB.java
@@ -1640,107 +1640,8 @@ public class ExamenEJB {
     
 
    
-   public List<Categoria> buscarSubCategoriaFrotis( )
-    {
-        Query query = em.createNamedQuery("Categoria.findByIdCategoria");
-        query.setParameter("idCategoria", 5);
-        return query.getResultList();
-    }
-   
-   public List<Categoria> buscarSubCategoriaObservacion( )
-    {
-        Query query = em.createNamedQuery("Categoria.findAll");
-        return query.getResultList();
-    }
-   
-   public List<Categoria> buscarSubCategoriaResultado( )
-    {
-        Query query = em.createNamedQuery("Categoria.findByIdCategoria");
-        query.setParameter("idCategoria", 6);
-        return query.getResultList();
-    }
-    
-   public List<SubCategoria> buscarValoresFrotis( BigDecimal codigo )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", codigo);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarValoresResultado( BigDecimal codigo )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", codigo);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarValorFuma( )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 7);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarValorToma( )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 8);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarValorEmbarazoActual( )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 9);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarObservaciones( )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 23);
-        return query.getResultList();
-    }
-   
-   
-   
-   
-   
-       //Codigo Angelo para Antecedentes Ginecobtetrico
-        
-   
-   
-   public List<Categoria> CargarTodasLasCategorias()
-    {
-        Query query = em.createNamedQuery("Categoria.findAll");
-        return query.getResultList();
-        
-    }
-   //davis parte
-   public List<SubCategoria> buscarProcedencia( )
-    {
-        
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 10);
-        return query.getResultList();
-    }
-   
-   public List<SubCategoria> buscarAspectoClinico( )
-    {
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 11);
-        return query.getResultList();
-    }
-    
-     public List<SubCategoria> buscarSecrecion( )
-    {
-        
-        
-        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
-        query.setParameter("idSubcategoria", 12);
-        return query.getResultList();
-    }
-   
+  
+  
      
      //codigo angelo recidencia
