+++ src/java/ni/gob/minsa/sivipcan/controlador/SubCategoriaEJB.java
@@ -5,6 +5,7 @@
  */
 package ni.gob.minsa.sivipcan.controlador;
 
+import java.math.BigDecimal;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
@@ -42,5 +43,86 @@ public class SubCategoriaEJB {
         subCategoria = em.find(SubCategoria.class, subCategoria.getIdSubcategoria());
         em.remove(subCategoria);
     }
+    
+    
+    
+    /**
+     * metodos para cargar listas y objetos de subcategoria para examen pap
+     * @param codigo
+     * @return 
+     */
+    
+    
+    
+    
+ 
+   public List<SubCategoria> buscarValoresFrotis( BigDecimal codigo )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", codigo);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarValoresResultado( BigDecimal codigo )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", codigo);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarValorFuma( )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 7);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarValorToma( )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 8);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarValorEmbarazoActual( )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 9);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarObservaciones( )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 23);
+        return query.getResultList();
+    }
+
+   public List<SubCategoria> buscarProcedencia( )
+    {
+        
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 10);
+        return query.getResultList();
+    }
+   
+   public List<SubCategoria> buscarAspectoClinico( )
+    {
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 11);
+        return query.getResultList();
+    }
+    
+     public List<SubCategoria> buscarSecrecion( )
+    {
+        
+        Query query = em.createNamedQuery("SubCategoria.findByIdSubcategoria");
+        query.setParameter("idSubcategoria", 12);
+        return query.getResultList();
+    }
+   
+    
+    
+    
  
 }
