+++ web/Valores.xhtml
@@ -42,7 +42,7 @@ and open the template in the editor.
                     </p:tabView>
 
                     <p:growl id="RegionGrowl" showDetail="false" autoUpdate="true" life="4000"/>
-                    <p:dataTable id="regionesTables" var="valores" value="#{valoresMB.listaValores}" rowKey="#{valores.idValor}"
+                    <p:dataTable id="regionesTables" var="valores" value="#{valoresMB.listaValores2}" rowKey="#{valores.idValor}"
                                  selection="#{valoresMB.seleccion}" selectionMode="single"
                                  paginator="true" rows="10"
                                  paginatorTemplate="{CurrentPageReport}  {FirstPageLink} {PreviousPageLink} {PageLinks} {NextPageLink} {LastPageLink} {RowsPerPageDropdown}"
